move single threaded ws benchmark into its own module
